+++ server/src/auth/auth.controller.spec.ts
@@ -1,0 +1,1 @@
+// https://github.com/leosuncin/nest-auth-example/blob/master/src/auth/auth.controller.spec.ts
