+++ server/src/auth/jwt.strategy.ts
